+++ tests/ByteDev.Json.SystemTextJson.UnitTests/Serialization/UnixEpochTimeToDateTimeJsonConverterTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using ByteDev.Json.SystemTextJson.Serialization;
+using ByteDev.Json.SystemTextJson.UnitTests.TestEntities;
 using NUnit.Framework;
 
 namespace ByteDev.Json.SystemTextJson.UnitTests.Serialization