+++ tests/ByteDev.Json.SystemTextJson.UnitTests/TestEntities/TestDateTimeEntity.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Text.Json.Serialization;
 
-namespace ByteDev.Json.SystemTextJson.UnitTests.Serialization
+namespace ByteDev.Json.SystemTextJson.UnitTests.TestEntities
 {
     public class TestDateTimeEntity
     {